Restructure the Python test suite.
